Adjustments to build scripts regarding use of $ROOT_DIR. Mostly using $WORKING_DIR now, and keeping the usage confined to a single script not across source calls. Now using git rev-parse to detect $ROOT_DIR of repo for sourcing of qflock_version.  We will also use for common in future. Split out setup_build.sh from the setup.sh. Changed remove of dockers to first check if dockers exist before removing to remove warnings.

+++ spark/docker/clean.sh
@@ -1,20 +1,25 @@
 #!/bin/bash
-pushd "$(dirname "$0")" # connect to root
-ROOT_DIR=$(pwd)
-echo "ROOT_DIR ${ROOT_DIR}"
+pushd "$(dirname "$0")"
 
 source setup.sh
 
 if [ "${NO_CLEAN_DOCKERS}" != "1" ]; then
     echo "Spark is cleaning dockers"
-    docker rmi -f $SPARK_DOCKER_BASE_NAME || true
-    docker rmi -f $SPARK_DOCKER_NAME || true
+    DOCKER_IMAGE_BASE=$(docker images -q $SPARK_DOCKER_BASE_NAME)
+    if [ ! -z $DOCKER_IMAGE_BASE ]; then
+        docker rmi -f $SPARK_DOCKER_BASE_NAME || true
+    fi
+    DOCKER_IMAGE=$(docker images -q $SPARK_DOCKER_NAME)
+    if [ ! -z $DOCKER_IMAGE ]; then
+        docker rmi -f $SPARK_DOCKER_NAME || true
+    fi
     docker system prune -f
 else
     echo "Spark is not cleaning dockers, NO_CLEAN_DOCKERS=1"
 fi
 
-if [ -e ${ROOT_DIR}/spark*.tgz ]; then
-    rm -rf ${ROOT_DIR}/spark*.tgz
+if [ -e ./spark*.tgz ]; then
+    rm -rf ./spark*.tgz
+    rm -rf ./hadoop*.tar.gz
 fi
 popd
